restructure with node modules

+++ tokens/olympus-dao/src/OlympusDaoTrap.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.12;
 
 import "./interfaces/IBondFixedExpiryTellerWithDrosera.sol";
-import {ERC20} from "../lib/solmate/src/tokens/ERC20.sol";
+import {ERC20} from "solmate/src/tokens/ERC20.sol";
 import {ITrap} from "drosera-contracts/interfaces/ITrap.sol";
 
 // Drosera Trap Contract
@@ -10,10 +10,9 @@ import {ITrap} from "drosera-contracts/interfaces/ITrap.sol";
 contract OlympusDaoTrap is ITrap {
     // State Variables
     // - Number of blocks between points
-    uint public blockInterval = 5;
+    uint256 public blockInterval = 5;
     address constant OHM = 0x64aa3364F17a4D01c6f1751Fd97C2BD3D7e7f1D5;
-    address constant treasuryAddress =
-        0x007FE7c498A2Cf30971ad8f2cbC36bd14Ac51156;
+    address constant treasuryAddress = 0x007FE7c498A2Cf30971ad8f2cbC36bd14Ac51156;
     address public protocol;
     address public account;
 
@@ -26,18 +25,15 @@ contract OlympusDaoTrap is ITrap {
 
     // Core Functions
     function collect() external view returns (bytes memory) {
-        uint treasury = ERC20(protocol).balanceOf(account);
-        uint[] memory result = new uint[](1);
+        uint256 treasury = ERC20(protocol).balanceOf(account);
+        uint256[] memory result = new uint256[](1);
         result[0] = treasury;
         return abi.encode(result);
     }
 
-    function shouldRespond(
-        bytes[] calldata data
-    ) external pure returns (bool, bytes memory) {
-        
-        uint[] memory currentX = abi.decode(data[0], (uint[]));
-        uint[] memory previousX = abi.decode(data[1], (uint[]));
+    function shouldRespond(bytes[] calldata data) external pure returns (bool, bytes memory) {
+        uint256[] memory currentX = abi.decode(data[0], (uint256[]));
+        uint256[] memory previousX = abi.decode(data[1], (uint256[]));
         if (previousX[0] > currentX[0]) {
             // Negative difference or no change
             if ((100 * (previousX[0] - currentX[0])) / previousX[0] <= 30) {
